Pet_adoption_project/streamlit/requirements.txt
streamlit==1.11.1
pycaret==2.3.10
pandas==1.4.3
numpy==1.23.1
opencv-python-headless==4.6.0.66
tensorflow==2.9.1
transformers==4.21.0
tokenizers==0.12.1
tqdm==4.64.0
torch==1.12.0
torchvision==0.13.0
scikit-learn==0.23.2
scipy==1.5.4
nltk==3.7
sentencepiece==0.1.96
huggingface-hub==0.8.1
emoji==2.0.0
pillow==9.2.0
sentence-transformers==2.2.2
numba==0.53.0